// frontend/src/features/actividades/registrar_escritura_cbr/models/registrar_escritura_cbr.ts
export interface RegistrarEscrituraCbr {
  id_registrar_escritura_cbr: number;
  id_expediente: number;
  conservador?: string | null;
  numero_caratula?: string | null;
  observaciones?: string | null;

  is_active: boolean;
  row_status: boolean;
  created_by: number;
  created_date: string;
  modified_by?: number | null;
  modified_date?: string | null;
}